Set the AST_INFERENCE_MODE to false
Since each preprocessor has it's own special way of
needing a combination of these flags, this approach
does not scale.

Stas used to have them in launch configurations, but
those were not versioned, so the information was lost
for mankind. The plan is to do some kind of passable
configuration so the thing can be parameterized in code
and therefore versionable.