Trying to deal with the abstracted node content
That is important in determining the "type" of a
transformation. Changin a public to a private is different
than changing a private to a public. I need to take this
into account when I have changes comming in.

One solution is to replay, and see if I can find the
affected node. Tricky, but it's slowly taking me
somewhere.